% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_normalization.R
\name{apply_dlfc_correction}
\alias{apply_dlfc_correction}
\title{wrapper function to apply several steps of correction to differential LFC scores from numerous screens combined}
\usage{
apply_dlfc_correction(
  scores,
  output_folder,
  control_dlfc_filepath,
  remove_screen_list = c(),
  drug_list_4x4 = c(),
  null_drug_list = c(),
  save_intermediate = FALSE
)
}
\arguments{
\item{drug_list_4x4}{A named nested list of 4x4 screens; used in the 'within correlation' calculation in wbc score.
example of one entry: "CHEM015_BORTEZOMIB_T14"=list("CHEM015_BORTEZOMIB_T14","CHEM050_BORTEZOMIB_T14","CHEM051_BORTEZOMIB_T14","CHEM052_BORTEZOMIB_T13").}

\item{null_drug_list}{A named nested list of all screens that are replicates of the 4x4 screens; these should have no effect in the 'between correlation' calculation in wbc score.
Example of one entry: "CHEM029_NGI1_T16"=list("CHEM031_NGI1_T13","CHEM050_NGI1_T11","CHEM029_NGI1_T16","CHEM050_NGI1_T14","CHEM051_NGI1_T14","CHEM057_NGI1_T13").
 @param save_intermediate If TRUE, save dLFC score file after each intermediate correction step (default FALSE).
 @return differential log fold change scores data-frame (library-genes X screens) after all batch correction steps.
 
 @export}
}
\description{
@param scores differential log fold change scores data-frame from condition screens (library-genes X screens, each screen column contain dLFC scores).
 @param output_folder path to directory to save all output files.
 @param control_dlfc_filepath path to control dLFC score file "control_effect_scores.tsv" (in folder 'control') generated by function generate_control_dlfc_scores(). 
 The path is hard-coded to reflect the generated file.
 @param remove_screen_list list of screens to remove from batch correction.
}
