% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_normalization.R
\name{apply_SD_scaling}
\alias{apply_SD_scaling}
\title{Apply standard-deviation scaling to dLFC score data-frame}
\usage{
apply_SD_scaling(scores, sd_table_output_directory)
}
\arguments{
\item{scores}{differential log fold change scores data-frame (library-genes X screens; each screen column contain dLFC scores)}

\item{sd_table_output_directory}{directory path to save the sd_table.tsv (contains per screen standard-deviations pre- and post- sd-scaling)}
}
\value{
differential log fold change scores data-frame (library-genes X screens) after applying SD scaling per screen (column)
}
\description{
Calculates per screen standard deviation.
Calculate target standard-deviations per screen i.e. sd of scores between 10%-90% percentiles.
Normalize target standard-deviations per screen by average of all target standard-deviations.
Normalize each screen score with target standard-deviation.
}
