% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_control_effect.R
\name{batch_control_cor_check}
\alias{batch_control_cor_check}
\title{Correlation check function for control replicates
Will pick the screens with poor control replicate correlations given a cutoff}
\usage{
batch_control_cor_check(
  df,
  screens,
  output_folder,
  pcc_cutoff = 0.5,
  verbose = TRUE,
  control_keyword = "DMSO|MOCK|Mock|WT|NGLY1|BMI1|Control"
)
}
\arguments{
\item{df}{Reads or lfc dataframe.}

\item{screens}{List of screens created with \code{add_screens}.}

\item{output_folder}{Folder for output.}

\item{pcc_cutoff}{Pearson correlation cutoff below which we define as poorly-correlated controls.}

\item{verbose}{If true, prints verbose output (default FALSE).}

\item{control_keyword}{Keyword (e.g. DMSO) in the control replicate name for identification.}
}
\value{
Two dataframes: first for screens passing the pcc cutoff, second for "bad" screens.
}
\description{
Correlation check function for control replicates
Will pick the screens with poor control replicate correlations given a cutoff
}
