% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_single_screen.R
\name{call_drug_hits}
\alias{call_drug_hits}
\title{Call significant responses for scored chemogenomic data.}
\usage{
call_drug_hits(
  scores,
  control_screen_name,
  condition_screen_names,
  fdr_threshold_positive = 0.1,
  fdr_threshold_negative = 0.1,
  differential_threshold_positive = 0,
  differential_threshold_negative = 0,
  neg_type = "Negative",
  pos_type = "Positive"
)
}
\arguments{
\item{scores}{Dataframe returned from \code{score_condition_vs_control}}

\item{control_screen_name}{Name of a control screen to test condition screens against}

\item{condition_screen_names}{A list of condition screen names to score against the 
control screen.}

\item{fdr_threshold_positive}{Threshold below which to call gene effects as significant positive hits
(default 0.1).}

\item{fdr_threshold_negative}{Threshold below which to call gene effects as significant negative hits
(default 0.1).}

\item{differential_threshold_positive}{Threshold on differential effects, 
over which to call gene effects as significant positive hits (default 0).}

\item{differential_threshold_negative}{Threshold on differential effects, 
below which gene effects are called as significant negative hits (default 0).}

\item{neg_type}{Label for significant effects with a negative differential effect
(default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
(default "Positive").}
}
\value{
Dataframe of scored data with differential effects called as significant
  for the specified conditions.
}
\description{
Run this to call significant responses for data returned from 
\code{score_condition_vs_control}.
}
