% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_control_effect.R
\name{compute_control_effects}
\alias{compute_control_effects}
\title{Computes effect sizes for control screens.}
\usage{
compute_control_effects(
  control_df,
  control_names,
  bad_cor_controls,
  loess = TRUE,
  ma_transform = TRUE
)
}
\arguments{
\item{control_df}{LFC dataframe for control screens.}

\item{control_names}{List of screen names for control screens to compute effect
sizes for.}

\item{bad_cor_controls}{A list of control screen names tested to have bad correlations.}

\item{loess}{If true, loess-normalizes residuals before running hypothesis testing.
Only works when test = "moderated-t" (default TRUE).}

\item{ma_transform}{If true, M-A transforms data before running loess normalization. Only
has an effect when loess = TRUE (default TRUE).}
}
\value{
Dataframe of effect sizes for control screens.
}
\description{
Computes rough estimates of effect sizes for control screens, scored against
all control screens passing QC with equal weights given to each.
}
