% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_control_effect.R
\name{compute_control_weights}
\alias{compute_control_weights}
\title{Computes weights for control screens.}
\usage{
compute_control_weights(control_names, condition_names, bad_cor_controls)
}
\arguments{
\item{control_names}{List of screen names for control screens to compute effect
sizes for.}

\item{condition_names}{List of screen names for control screens to compute effect
sizes for.}

\item{bad_cor_controls}{A list of control screen names tested to have bad correlations.}
}
\value{
Dataframe of weights for each condition and control with the number of 
  rows equal to length(condition_names) and the number of columns equal to 
  length(control_names).
}
\description{
Computes weights for control screens based on matching condition screens
for \code{score_controls_vs_controls}.
}
