% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_normalization.R
\name{generate_control_dlfc_scores}
\alias{generate_control_dlfc_scores}
\title{wrapper function to generate differential log fold change from control screens}
\usage{
generate_control_dlfc_scores(
  condition_control_map_file,
  screen_replicate_map_file,
  raw_read_count_file,
  output_folder,
  filter_names_postfix = "T0",
  cf1 = 1e+06,
  cf2 = 1,
  min_reads = 30,
  max_reads = 10000,
  nonessential_norm = TRUE,
  replace_NA = TRUE,
  min_guides = 3,
  loess = TRUE,
  ma_transform = TRUE,
  control_genes = c("None", ""),
  qc_control_pcc = 0.3,
  verbose = FALSE,
  black_list = c("bad_screen_name", ""),
  screen_control_keyword = "DMSO|MOCK|Mock|WT|NGLY1|BMI1|Control"
)
}
\arguments{
\item{filter_names_postfix}{Postfix to identify list of screen names to filter based on read counts by.}

\item{cf1}{parameter for \code{normalize_screens}. Scaling factor (default 1e6).}

\item{cf2}{parameter for \code{normalize_screens}. Pseudocount (default 1).}

\item{min_reads}{parameter for \code{normalize_screens}. Minimum number of reads to keep (default 30, anything
below this value will be filtered out).}

\item{max_reads}{parameter for \code{normalize_screens}. Maximum number of reads to keep (default 10000, anything
above this value will be filtered out).}

\item{nonessential_norm}{parameter for \code{normalize_screens}. Whether or not to normalize each screen against its
population of core non-essential genes, as defined by Traver et al. 2015 
(default FALSE).}

\item{replace_NA}{parameter for \code{normalize_screens}.  Whether or not to replace NA and NULL values in non-T0 screens 
with 0's after filtering out T0 guides with too few or NA readcounts 
(default TRUE).}

\item{min_guides}{parameter for \code{score_controls_batch}. The minimum number of guides per gene pair required to score data (default 3).}

\item{loess}{parameter for \code{score_controls_batch}. If true, loess-normalizes residuals before running hypothesis testing. (default TRUE).}

\item{ma_transform}{parameter for \code{score_controls_batch}. If true, M-A transforms data before running loess normalization. Only
has an effect when loess = TRUE (default TRUE).}

\item{control_genes}{parameter for \code{score_controls_batch}. List of control genes to remove, e.g. "luciferase" (default c("None", "")).}

\item{qc_control_pcc}{parameter for \code{score_controls_batch}. (default 0.30).}

\item{verbose}{parameter for \code{score_controls_batch}. If true, prints verbose output (default FALSE).}

\item{black_list}{parameter for \code{score_controls_batch}. A list of control screens to be excluded beforehand (default c('bad_screen_name', "")).}

\item{screen_control_keyword}{parameter for \code{score_controls_batch}. Key words to identify control screen names (default "DMSO|MOCK|Mock|WT|NGLY1|BMI1|Control")
Output: creates file dLFC score file "control_effect_scores.tsv" in folder 'control' 
 
 @export}
}
\description{
@param condition_control_map_file Path to .tsv file mapping screens to their controls for scoring, with two 
  columns for "Screen" and "Control." 
 @param screen_replicate_map_file Path to .tsv file mapping screens to their replicates and earlier time point screens,
with three columns 'Screen','Replicates', and 'NormalizeTo'
 @param raw_read_count_file path to raw read-count file containing the control screen replicates.
 @param output_folder  Path to folder where new output files will be stored. A folder named 'control' will be created in the output_folder.
}
