% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_single_screen.R
\name{get_screen_batch_name}
\alias{get_screen_batch_name}
\title{Gets the screen-batch name from the screen name
Example: screen name: CHEM001_BORTEZOMIB_T18;  from this returns CHEM001 which is the screen-batch name}
\usage{
get_screen_batch_name(x)
}
\arguments{
\item{x}{string containing the screen name in the format screen-batch-name_compound-name_time}
}
\value{
string containing the screen-batch name
}
\description{
Gets the screen-batch name from the screen name
Example: screen name: CHEM001_BORTEZOMIB_T18;  from this returns CHEM001 which is the screen-batch name
}
