% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_single_screen.R
\name{jackknife_outliers}
\alias{jackknife_outliers}
\title{Finds and removes (set to NA) outlier guides from the conditional residual (differential scores) for each gene}
\usage{
jackknife_outliers(cond_res, threshold = 2)
}
\arguments{
\item{cond_res}{dataframe of condition residuals (differential scores); rows: genes, column: replicate-guide pair}

\item{threshold}{jack-knife threshold to decide if a guide is an outlier (default value 2)}
}
\value{
dataframe of updated condition residuals (differential scores); rows: genes, column: replicate-guide pair
}
\description{
Finds and removes (set to NA) outlier guides from the conditional residual (differential scores) for each gene
}
