% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_guide_level_dlfc_for_hits}
\alias{plot_guide_level_dlfc_for_hits}
\title{Plot guide-level differential LFC scores for all hits}
\usage{
plot_guide_level_dlfc_for_hits(
  scores,
  residuals,
  control_name,
  condition_name,
  output_folder,
  neg_type = "Negative",
  pos_type = "Positive",
  plot_type = "png"
)
}
\arguments{
\item{scores}{Dataframe of various scores - loaded from the file written after score_drugs_batch() call}

\item{residuals}{Dataframe guide-level differential LFCs  (loaded from the file written after score_drugs_batch() call with the save_guide_dlfc argument set to TRUE)}

\item{control_name}{Name of control}

\item{condition_name}{Name of condition}

\item{output_folder}{Folder to output plots to}

\item{neg_type}{Label for significant effects with a negative differential effect (default "Negative", should match entries in the scores dataframe).}

\item{pos_type}{Label for significant effects with a positive differential effect (default "Positive", should match entries in the scores dataframe).}

\item{plot_type}{Type of plot to output, one of "png" or "pdf" (default "png").}
}
\description{
Plots guide-level differential LFC scores for each called hit and outputs plots to a given folder.
}
