% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_normalization.R
\name{remove_control_dlfc_signal}
\alias{remove_control_dlfc_signal}
\title{remove signal from control screens from dLFC scores}
\usage{
remove_control_dlfc_signal(scores, control_dlfc_filepath)
}
\arguments{
\item{scores}{differential log fold change scores data-frame from condition screens (library-genes X screens, each screen column contain dLFC scores)}

\item{control_dlfc_filepath}{file path to DMSO (control screens) differential log fold change score file (generated by OROBAS pipeline )}
}
\value{
differential log fold change scores data-frame (library-genes X screens) after removing principal components
}
\description{
load differential log fold change scores generated from control screens from mentioned file.
apply principal component analysis to the control dLFC scores.
create projection of all principal components from control-dLFCs onto the main(condition) dLFC score matrix.
subtract the the projection matrix from condition score matrix to remove dLFC signal of control screens
}
