% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_normalization.R
\name{remove_principal_component_signal}
\alias{remove_principal_component_signal}
\title{remove principal component signal from dLFC scores}
\usage{
remove_principal_component_signal(scores, pc = 1)
}
\arguments{
\item{scores}{differential log fold change scores data-frame (library-genes X screens; each screen column contain dLFC scores)}

\item{pc}{number of principal components to remove; default 1}
}
\value{
differential log fold change scores data-frame (library-genes X screens) after removing principal components
}
\description{
apply principal component analysis on the dLFC score matrix.
calculate projection of first mentioned number of principal component onto the score matrix.
subtract the projection from dLFC score matrix to remove principal component signal.
}
