% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_normalization.R
\name{remove_screens}
\alias{remove_screens}
\title{remove screens not to be included in the dLFC score data-frame}
\usage{
remove_screens(scores, exclude_screen_list = c())
}
\arguments{
\item{scores}{differential log fold change scores data-frame (library-genes X screens; each screen column contain dLFC scores)}

\item{exclude_screen_list}{a list of screens to be removed from the scores data-frame; default is empty list c().}
}
\value{
differential log fold change scores data-frame (library-genes X screens) after removing selected screens (columns)
}
\description{
The scores matrix has library genes on the row side and compound screens on the column side.
A list of screens can be provided to be excluded and those columns will be removed from the scores matrix.
}
