% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_normalization.R
\name{run_global_normalization}
\alias{run_global_normalization}
\title{wrapper function to generate differential log fold change from control screens}
\usage{
run_global_normalization(
  input_folder,
  condition_control_map_file,
  output_folder,
  screen_replicate_map_file,
  raw_read_count_file,
  filter_names_postfix = "T0",
  cf1 = 1e+06,
  cf2 = 1,
  min_reads = 30,
  max_reads = 10000,
  nonessential_norm = TRUE,
  replace_NA = TRUE,
  min_guides = 3,
  loess = TRUE,
  ma_transform = TRUE,
  control_genes = c("None", ""),
  qc_control_pcc = 0.3,
  verbose = FALSE,
  black_list = c("bad_screen_name", ""),
  screen_control_keyword = "DMSO|MOCK|Mock|WT|NGLY1|BMI1|Control",
  remove_screen_list = c(),
  drug_list_4x4 = c(),
  null_drug_list = c(),
  save_intermediate = FALSE,
  neg_type = "Negative",
  pos_type = "Positive",
  fdr_threshold_positive = 0.1,
  fdr_threshold_negative = 0.1,
  differential_threshold_positive = 0,
  differential_threshold_negative = 0,
  plot_type = "png",
  label_fdr_threshold = NULL
)
}
\arguments{
\item{filter_names_postfix}{parameter for \code{generate_control_dlfc_scores}.}

\item{cf1}{parameter for \code{generate_control_dlfc_scores}.  Scaling factor (default 1e6).}

\item{cf2}{parameter for \code{generate_control_dlfc_scores}. Pseudocount (default 1).}

\item{min_reads}{parameter for \code{generate_control_dlfc_scores}. Minimum number of reads to keep (default 30, anything
below this value will be filtered out).}

\item{max_reads}{parameter for \code{generate_control_dlfc_scores}. Maximum number of reads to keep (default 10000, anything
above this value will be filtered out).}

\item{nonessential_norm}{parameter for \code{generate_control_dlfc_scores}. Whether or not to normalize each screen against its
population of core non-essential genes, as defined by Traver et al. 2015 
(default FALSE).}

\item{replace_NA}{parameter for \code{generate_control_dlfc_scores}. Whether or not to replace NA and NULL values in non-T0 screens 
with 0's after filtering out T0 guides with too few or NA readcounts 
(default TRUE).}

\item{min_guides}{parameter for \code{generate_control_dlfc_scores}. The minimum number of guides per gene pair required to score data (default 3).}

\item{loess}{parameter for \code{generate_control_dlfc_scores}. If true, loess-normalizes residuals before running hypothesis testing. (default TRUE).}

\item{ma_transform}{parameter for \code{generate_control_dlfc_scores}. If true, M-A transforms data before running loess normalization. Only
has an effect when loess = TRUE (default TRUE).}

\item{control_genes}{parameter for \code{generate_control_dlfc_scores}. List of control genes to remove, e.g. "luciferase" (default c("None", "")).}

\item{qc_control_pcc}{parameter for \code{generate_control_dlfc_scores}. (default 0.30).}

\item{verbose}{parameter for \code{generate_control_dlfc_scores}. If true, prints verbose output (default FALSE).}

\item{black_list}{parameter for \code{generate_control_dlfc_scores}. A list of control screens to be excluded beforehand (default c('bad_screen_name', "")).}

\item{screen_control_keyword}{parameter for \code{generate_control_dlfc_scores}. Key words to identify control screen names (default "DMSO|MOCK|Mock|WT|NGLY1|BMI1|Control")}

\item{remove_screen_list}{parameter for \code{apply_dlfc_correction}. List of screens to remove from batch correction.}

\item{drug_list_4x4}{parameter for \code{apply_dlfc_correction}. A named nested list of 4x4 screens; used in the 'within correlation' calculation in wbc score.
example of one entry: "CHEM015_BORTEZOMIB_T14"=list("CHEM015_BORTEZOMIB_T14","CHEM050_BORTEZOMIB_T14","CHEM051_BORTEZOMIB_T14","CHEM052_BORTEZOMIB_T13").}

\item{null_drug_list}{parameter for \code{apply_dlfc_correction}. A named nested list of all screens that are replicates of the 4x4 screens; these should have no effect in the 'between correlation' calculation in wbc score.
Example of one entry: "CHEM029_NGI1_T16"=list("CHEM031_NGI1_T13","CHEM050_NGI1_T11","CHEM029_NGI1_T16","CHEM050_NGI1_T14","CHEM051_NGI1_T14","CHEM057_NGI1_T13").}

\item{save_intermediate}{parameter for \code{apply_dlfc_correction}. If TRUE, save dLFC score file after each intermediate correction step (default FALSE).}

\item{neg_type}{parameter for \code{plot_drug_response} and \code{call_drug_hits}. Label for significant effects with a negative differential effect
passed to \code{call_drug_hits} (default "Negative").}

\item{pos_type}{parameter for\code{plot_drug_response} and \code{call_drug_hits}. Label for significant effects with a positive differential effect
passed to \code{call_drug_hits} (default "Positive").}

\item{fdr_threshold_positive}{parameter for \code{call_drug_hits}. The threshold below which to call gene effects as significant positive hits
(default 0.1).}

\item{fdr_threshold_negative}{parameter for \code{call_drug_hits}. The threshold below which to call gene effects as significant negative hits
(default 0.1).}

\item{differential_threshold_positive}{parameter for \code{call_drug_hits}. Threshold on differential effects, 
over which to call gene effects as significant positive hits (default 0).}

\item{differential_threshold_negative}{parameter for \code{call_drug_hits}. Threshold on differential effects, 
below which gene effects are called as significant negative hits (default 0).}

\item{plot_type}{parameter for \code{plot_drug_response}. Type of plot to output, one of "png" or "pdf" (default "png").}

\item{label_fdr_threshold}{parameter for \code{plot_drug_response}. The threshold below which to plot gene labels for significant
  hits, or NULL to plot without labels (default NULL).

Output: generates corrected dLFC score file and other files 
 
 @export}
}
\description{
@param input_folder Path to directory where all outputs from single screen scoring generated by running \code{run_single_screen_scoring} are stored.
 @param condition_control_map_file Path to .tsv file mapping screens to their controls for scoring, with two 
  columns for "Screen" and "Control." 
 @param screen_replicate_map_file Path to .tsv file mapping screens to their replicates and earlier time point screens,
with three columns 'Screen','Replicates', and 'NormalizeTo' 
 @param raw_read_count_file path to raw read-count file containing the control screen replicates.
 @param output_folder Path to folder where new output files will be stored. A folder named global_normalization will be created in the output_folder.
}
