% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_single_screen.R
\name{run_single_screen_scoring}
\alias{run_single_screen_scoring}
\title{Wrapper function to score screens and generate various plots.}
\usage{
run_single_screen_scoring(
  parent_folder,
  raw_read_count_file,
  screen_replicate_map_file,
  condition_control_map_file,
  plot_type = "png",
  display_numbers = TRUE,
  show_colnames = TRUE,
  show_rownames = TRUE,
  filter_names_postfix = "T0",
  cf1 = 1e+06,
  cf2 = 1,
  min_reads = 30,
  max_reads = 10000,
  nonessential_norm = TRUE,
  replace_NA = TRUE,
  min_guides = 3,
  loess = TRUE,
  ma_transform = TRUE,
  control_genes = c("None", ""),
  fdr_method = "BY",
  fdr_threshold_positive = 0.1,
  fdr_threshold_negative = 0.1,
  differential_threshold_positive = 0,
  differential_threshold_negative = 0,
  neg_type = "Negative",
  pos_type = "Positive",
  label_fdr_threshold = NULL,
  save_guide_dlfc = FALSE
)
}
\arguments{
\item{parent_folder}{Folder to output scored data and plots.}

\item{raw_read_count_file}{path to raw read-count file.}

\item{screen_replicate_map_file}{Path to .tsv file mapping screens to their repicates and earlier time point screen,
with three columns 'Screen','Replicates', and 'NormalizeTo'}

\item{condition_control_map_file}{Path to .tsv file mapping screens to their controls for scoring, with two 
columns for "Screen" and "Control."}

\item{plot_type}{Type of plot to output, one of "png" or "pdf" (default "png").}

\item{display_numbers}{Whether or not to include PCC values in heatmap (default TRUE).}

\item{show_colnames}{Whether or not to show column names on the plot (default TRUE).}

\item{show_rownames}{Whether or not to show row names on the plot (default TRUE).}

\item{filter_names_postfix}{substring to identify list of screen names to filter based on read counts by in the normalization step. (default 'T0')}

\item{cf1}{Scaling factor (default 1e6).}

\item{cf2}{Pseudocount (default 1).}

\item{min_reads}{Minimum number of reads to keep (default 30, anything
below this value will be filtered out).}

\item{max_reads}{Maximum number of reads to keep (default 10000, anything
above this value will be filtered out).}

\item{nonessential_norm}{Whether or not to normalize each screen against its
population of core non-essential genes, as defined by Traver et al. 2015 
(default FALSE).}

\item{replace_NA}{Whether or not to replace NA and NULL values in non-T0 screens 
with 0's after filtering out T0 guides with too few or NA readcounts 
(default TRUE).}

\item{min_guides}{The minimum number of guides per gene pair required to score data 
(default 3).}

\item{loess}{If true, loess-normalizes residuals before running hypothesis testing.
Only works when test = "moderated-t" (default TRUE).}

\item{ma_transform}{If true, M-A transforms data before running loess normalization. Only
has an effect when loess = TRUE (default TRUE).}

\item{control_genes}{List of control genes to remove, e.g. "luciferase" (default c("None", "")).}

\item{fdr_method}{Type of FDR to compute. One of "BH", "BY" or "bonferroni" (default
"BY")}

\item{fdr_threshold_positive}{Threshold below which to call gene effects as significant positive hits
(default 0.1).}

\item{fdr_threshold_negative}{Threshold below which to call gene effects as significant negative hits
(default 0.1).}

\item{differential_threshold_positive}{Threshold on differential effects, 
over which to call gene effects as significant positive hits (default 0).}

\item{differential_threshold_negative}{Threshold on differential effects, 
below which gene effects are called as significant negative hits (default 0).}

\item{neg_type}{Label for significant effects with a negative differential effect
(default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
(default "Positive").}

\item{label_fdr_threshold}{Threshold below which to plot gene labels for significant
hits, or NULL to plot without labels (default NULL).}

\item{save_guide_dlfc}{If true, saves guide-level dLFC scores for each condition screen to the output folder
(default FALSE).}
}
\description{
Wrapper function to score screens and generate various plots.
}
