% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_single_screen.R
\name{score_condition_vs_control}
\alias{score_condition_vs_control}
\title{Scores conditions against a single control.
single screen score function with individual replicate-level loess fitting (loess smoothing Drug A vs DMSO A, B vs B, C vs C per guide)}
\usage{
score_condition_vs_control(
  df,
  screens,
  control_screen_name,
  condition_screen_names,
  control_genes = c("None", ""),
  min_guides = 3,
  loess = TRUE,
  ma_transform = TRUE,
  fdr_method = "BY",
  return_residuals = FALSE
)
}
\arguments{
\item{df}{LFC dataframe.}

\item{screens}{List of screens generated with \code{add_screens}.}

\item{control_screen_name}{Name of a control screen to test condition screens against.}

\item{condition_screen_names}{Name of a condition screen to score against the 
control screen.}

\item{control_genes}{List of control genes to remove, e.g. "luciferase" (default c("None", "")).}

\item{min_guides}{The minimum number of guides per gene pair required to score data 
(default 3).}

\item{loess}{If true, loess-normalizes residuals before running hypothesis testing.}

\item{ma_transform}{If true, M-A transforms data before running loess normalization. Only
has an effect when loess = TRUE (default TRUE).}

\item{fdr_method}{Type of FDR to compute. One of "BH", "BY" or "bonferroni" (default "BY").}

\item{return_residuals}{If FALSE, returns NA instead of residuals dataframe (default FALSE).
This is recommend if scoring large datasets and memory is a limitation.}
}
\value{
A named list containing 3 dataframes. 
  i) "scored_data": a dataframe containing scored data with separate columns given by the specified control and condition names.
  ii) "guide_dlfc_pre_jk": a dataframe containing guide-level and replicate-level differential LFCs scored for condition screen against control screen, before applying jk-outlier removal.
}
\description{
Scores guides for any number of drug screens against a control screen
(e.g. for directly comparing drug response to DMSO response). After running 
this function, pass the resulting dataframe to \code{call_drug_hits} to 
call significant effects.
}
