% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_control_effect.R
\name{score_controls_batch}
\alias{score_controls_batch}
\title{Scores multiple control screens against consensus controls}
\usage{
score_controls_batch(
  df,
  screens,
  batch_file,
  output_folder,
  min_guides = 3,
  loess = TRUE,
  ma_transform = TRUE,
  control_genes = c("None", ""),
  qc_control_pcc = 0.3,
  verbose = FALSE,
  black_list = c("bad_screen_name", ""),
  screen_control_keyword = "DMSO|MOCK|Mock|WT|NGLY1|BMI1|Control"
)
}
\arguments{
\item{df}{LFC dataframe.}

\item{screens}{List of screens generated with \code{add_screens}.}

\item{batch_file}{Path to .tsv file mapping screens to their controls for scoring, with four 
columns for "Screen", "Control", "Group", and "Type".}

\item{output_folder}{Folder to output scored data and plots to.}

\item{min_guides}{The minimum number of guides per gene pair required to score data 
(default 3).}

\item{loess}{If true, loess-normalizes residuals before running hypothesis testing.
Only works when test = "moderated-t" (default TRUE).}

\item{ma_transform}{If true, M-A transforms data before running loess normalization. Only
has an effect when loess = TRUE (default TRUE).}

\item{control_genes}{List of control genes to remove, e.g. "luciferase" (default c("None", "")).}

\item{verbose}{If true, prints verbose output (default FALSE).}

\item{black_list}{A list of control screens to be excluded beforehand.}

\item{screen_control_keyword}{Key words to identify control screen names.}

\item{label_fdr_threshold}{Threshold below which to plot gene labels for significant
hits, or NULL to plot without labels (default NULL).}

\item{save_residuals}{If true, saves residuals for each screen to the output folder
(default FALSE).}
}
\description{
Checks whether the controls' average internal correlation pass the threshold (e.g. pcc=0.50),
based on which divide the screens into "good" and "bad".
The "good" group will be averaged into a consensus collection,
and all screens will be scored against average controls from all good screens.
}
