% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_control_effect.R
\name{score_controls_vs_controls}
\alias{score_controls_vs_controls}
\title{Scores conditions (=controls) against consensus controls.}
\usage{
score_controls_vs_controls(
  df,
  screens,
  control_screen_names,
  bad_cor_controls,
  control_genes = c("None", ""),
  min_guides = 2,
  loess = TRUE,
  ma_transform = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df}{LFC dataframe.}

\item{screens}{List of screens generated with \code{add_screens}.}

\item{control_screen_names}{A list of control screen names to test the condition
screens against.}

\item{bad_cor_controls}{A list of control screen names tested to have bad correlations.}

\item{control_genes}{List of control genes to remove, e.g. "luciferase" (default c("None", "")).}

\item{min_guides}{The minimum number of guides per gene pair required to score data 
(default 3).}

\item{loess}{If true, loess-normalizes residuals before running hypothesis testing.
Only works when test = "moderated-t" (default TRUE).}

\item{ma_transform}{If true, M-A transforms data before running loess normalization. Only
has an effect when loess = TRUE (default TRUE).}

\item{verbose}{If true, prints verbose output (default FALSE).}
}
\value{
A dataframe.
  Contains scored control effect data with separate columns for specified control
  names. It was aggregated from loess-normalized residuals for all guides.
}
\description{
Scores guides for any number of screens against multiple control screens
(e.g. for directly comparing control response to consensus control response).
}
