% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_drugs_vs_controls}
\alias{score_drugs_vs_controls}
\title{Scores conditions against multiple controls.}
\usage{
score_drugs_vs_controls(
  df,
  screens,
  control_screen_names,
  condition_screen_names,
  matched_controls,
  output_folder,
  control_genes = c("None", ""),
  min_guides = 2,
  loess = TRUE,
  fdr_method = "BY",
  n_components = 0,
  chromosomal_correction = FALSE,
  return_residuals = TRUE,
  intermediate_file = NULL,
  load_intermediate = FALSE,
  plot_type = "png",
  verbose = FALSE
)
}
\arguments{
\item{df}{LFC dataframe.}

\item{screens}{List of screens generated with \code{add_screens}.}

\item{control_screen_names}{A list of control screen names to test the condition
screens against.}

\item{condition_screen_names}{A list of condition screen names to score against the 
control screen.}

\item{matched_controls}{A list of control screen names with each screen corresponding
to the condition at the same position in condition_screen_names.}

\item{output_folder}{Folder to output scree plot to.}

\item{control_genes}{List of control genes to remove, e.g. "luciferase" (default c("None", "")).}

\item{min_guides}{The minimum number of guides per gene pair required to score data 
(default 3).}

\item{loess}{If true, loess-normalizes residuals before running hypothesis testing.
Only works when test = "moderated-t" (default TRUE).}

\item{fdr_method}{Type of FDR to compute. One of "BH", "BY" or "bonferroni" (default "BY").}

\item{n_components}{Number of principal components to remove from data.}

\item{chromosomal_correction}{If TRUE, corrects chromosomal shifts by down-weighting qGI
scores for shifted regions (default FALSE).}

\item{return_residuals}{If FALSE, returns NA instead of residuals dataframe (default TRUE).
This is recommend if scoring large datasets and memory is a limitation.}

\item{intermediate_file}{Path to file to save intermediate output to, or NULL to not write
any intermediate output (default NULL).}

\item{load_intermediate}{If TRUE, loads data from intermediate file instead of running 
earlier steps in the scoring pipeline (default FALSE).}

\item{plot_type}{Type of plot to output, one of "png" or "pdf" (default "png").}

\item{verbose}{If true, prints verbose output (default FALSE).}
}
\value{
A list containing two dataframes. The first entry, named "scored_data" in the list,
  contains scored data with separate columns given by the specified control and condition
  names. The second entry, named "residuals" in the list, is a dataframe containing control,
  condition and loess-normalized residuals for all guides.
}
\description{
Scores guides for any number of drug screens against multiple control screens
(e.g. for directly comparing drug response to DMSO response). After running 
this function, pass the resulting dataframe to \code{call_drug_hits} to 
call significant effects.
}
