% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_single_screen.R
\name{score_single_screen}
\alias{score_single_screen}
\title{Scores multiple drugs against multiple controls}
\usage{
score_single_screen(
  df,
  screens,
  batch,
  output_folder,
  min_guides = 3,
  loess = TRUE,
  ma_transform = TRUE,
  control_genes = c("None", ""),
  fdr_method = "BY",
  fdr_threshold_positive = 0.1,
  fdr_threshold_negative = 0.1,
  differential_threshold_positive = 0,
  differential_threshold_negative = 0,
  neg_type = "Negative",
  pos_type = "Positive",
  label_fdr_threshold = NULL,
  save_guide_dlfc = FALSE,
  plot_type = "png"
)
}
\arguments{
\item{df}{LFC dataframe.}

\item{screens}{List of screens generated with \code{add_screens}.}

\item{batch}{dataframe mapping screens to their controls for scoring, with two 
columns for "Screen" and "Control."}

\item{output_folder}{Folder to output scored data and plots to.}

\item{min_guides}{The minimum number of guides per gene pair required to score data 
(default 3).}

\item{loess}{If true, loess-normalizes residuals before running hypothesis testing.
Only works when test = "moderated-t" (default TRUE).}

\item{ma_transform}{If true, M-A transforms data before running loess normalization. Only
has an effect when loess = TRUE (default TRUE).}

\item{control_genes}{List of control genes to remove, e.g. "luciferase" (default c("None", "")).}

\item{fdr_method}{Type of FDR to compute. One of "BH", "BY" or "bonferroni" (default
"BY")}

\item{fdr_threshold_positive}{Threshold below which to call gene effects as significant positive hits
(default 0.1).}

\item{fdr_threshold_negative}{Threshold below which to call gene effects as significant negative hits
(default 0.1).}

\item{differential_threshold_positive}{Threshold on differential effects, 
over which to call gene effects as significant positive hits (default 0).}

\item{differential_threshold_negative}{Threshold on differential effects, 
below which gene effects are called as significant negative hits (default 0).}

\item{neg_type}{Label for significant effects with a negative differential effect
(default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
(default "Positive").}

\item{label_fdr_threshold}{Threshold below which to plot gene labels for significant
hits, or NULL to plot without labels (default NULL).}

\item{save_guide_dlfc}{If true, saves guide-level dLFC scores for each condition screen to the output folder
(default FALSE).}

\item{plot_type}{Type of plot to output, one of "png" or "pdf" (default "png").}
}
\description{
Takes in an input .tsv file with two columns for "Screen" and "Control" and scores
all screens listed in "Screen" against their corresponding screens listed in
"Control." Outputs all files and plots in the specified folder.
}
