% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_normalization.R
\name{score_wbc}
\alias{score_wbc}
\title{generate wbc (within-between correlation) score of 4x4 screens}
\usage{
score_wbc(data, drug_list_4x4, null_drug_list)
}
\arguments{
\item{data}{differential log fold change scores data-frame (library-genes X screens, each screen column contain dLFC scores)}

\item{drug_list_4x4}{A named nested list of 4x4 screens; used in the 'within correlation' calculation in wbc score
example of one entry: "CHEM015_BORTEZOMIB_T14"=list("CHEM015_BORTEZOMIB_T14","CHEM050_BORTEZOMIB_T14","CHEM051_BORTEZOMIB_T14","CHEM052_BORTEZOMIB_T13")}

\item{null_drug_list}{A named nested list of all screens that are replicates of the 4x4 screens; these should have no effect in the 'between correlation' calculation in wbc score
Example of one entry: "CHEM029_NGI1_T16"=list("CHEM031_NGI1_T13","CHEM050_NGI1_T11","CHEM029_NGI1_T16","CHEM050_NGI1_T14","CHEM051_NGI1_T14","CHEM057_NGI1_T13")}
}
\value{
data-frame with one wbc score for each 4x4 screens
}
\description{
Score representing how much the replicate screens are correlated compared to non-replicate screens
Equation: wbc_per_screen = (average(within_correlation) - average(between_correlation))/sig_b
         within_correlation = pearson correlation of screen with other chosen replicates
         between_correlation = pearson_correlation of screen with all non-replicate screens
         sig_b = sqrt(sum( square(between_correlation - average(between_correlation)) ))/(number of non-replicate screens - 1))
<cite paper>
}
