% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_normalization.R
\name{screen_batch_correction_with_lda}
\alias{screen_batch_correction_with_lda}
\title{wrapper function to call python script screen_batch_correction_LDA.py}
\usage{
screen_batch_correction_with_lda(scores, output_directory)
}
\description{
@param scores differential log fold change scores data-frame from condition screens (library-genes X screens, each screen column contain dLFC scores)
 @param output_directory path to directory to save output files (roc plot files)
 @return differential log fold change scores data-frame (library-genes X screens) after batch correction by LDA
 
 @export
}
