% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{essential_lfc_qc}
\alias{essential_lfc_qc}
\title{Computes essential gene recovery AUC.}
\usage{
essential_lfc_qc(df, screens, gene_col, output_folder)
}
\arguments{
\item{df}{LFC dataframe.}

\item{screens}{List of screens generated with \code{add_screens}.}

\item{gene_col}{A column containing gene names for all guides.}

\item{output_folder}{Folder to which essential gene QC results should be saved.}
}
\value{
All output is saved to the file "essential_PR_QC.txt" in the folder given by
  output_folder.
}
\description{
Computes area under the curve for ROC curves that measure how well each technical replicate
recovers signal for essential-targeting guides and saves results to file.
}
